Fixed newsfeed sizing bug, added function to toggle articles from grid/list layout, styled button, fixed profile button sizing, fixed homepage button sizing cascade
