+++ src/styles/app.scss
@@ -9,6 +9,7 @@
 @import './search.scss';
 @import './friendRequests.scss';
 @import './squadPage.scss';
+@import './feed.scss';
 
 $theme-colors: (
   'primary': #0270f2,
