+++ src/styles/container.scss
@@ -1,0 +1,6 @@
+.squadPage {
+  padding-left: 2vw;
+  padding-right: 2vw;
+  padding-top: 2vh;
+
+}
